+++ resources/views/components/footer-copy.blade.php
@@ -1,7 +1,8 @@
 @php
 $serviceLink = sprintf(
-    '<a href="https://%s">%s</a>',
+    '<a href="https://%s" data-testid="%s">%s</a>',
     __('layouts.provider.www'),
+    'provider-link',
     __('layouts.provider.name')
 );
 @endphp