+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Add `data-testid` attributes to components for easier testing.
+
 ## [2.1.1] - 2025-05-31
 
 ### Fixed
@@ -506,6 +512,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [2.1.1]: https://https://github.com/internetguru/laravel-common/compare/v2.1.0...v2.1.1
 [2.1.0]: https://https://github.com/internetguru/laravel-common/compare/v2.0.5...v2.1.0
 [2.1.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v2.0.5